+++ cdp-agentkit-core/cdp_agentkit_core/actions/wow/create_token.py
@@ -11,7 +11,7 @@ from cdp_agentkit_core.actions.wow.constants import (
 )
 
 WOW_CREATE_TOKEN_PROMPT = """
-This tool will create a Zora Wow ERC20 memecoin using the WoW factory. This tool takes the token name and token symbol. It uses a bonding curve so there is no need to add liquidity to the pool upfront. It is only supported on Base Sepolia and Base Mainnet.
+This tool will create a Zora Wow ERC20 memecoin using the WoW factory. This tool takes the token name, token symbol, and optionally a token URI containing metadata about the token. It uses a bonding curve so there is no need to add liquidity to the pool upfront. It is only supported on Base Sepolia and Base Mainnet.
 """
 
 
@@ -26,15 +26,20 @@ class WowCreateTokenInput(BaseModel):
         ...,
         description="The symbol of the token to create, e.g. WOW",
     )
+    token_uri: str = Field(
+        default=None,
+        description="The URI of the token metadata to store on IPFS, e.g. ipfs://QmY1GqprFYvojCcUEKgqHeDj9uhZD9jmYGrQTfA9vAE78J",
+    )
 
 
-def wow_create_token(wallet: Wallet, name: str, symbol: str) -> str:
+def wow_create_token(wallet: Wallet, name: str, symbol: str, token_uri: str | None = None) -> str:
     """Create a Zora Wow ERC20 memecoin.
 
     Args:
         wallet (Wallet): The wallet to create the token from.
         name (str): The name of the token to create.
         symbol (str): The symbol of the token to create.
+        token_uri (str | None): The URI of the token metadata to store on IPFS e.g. ipfs://QmY1GqprFYvojCcUEKgqHeDj9uhZD9jmYGrQTfA9vAE78J.
 
     Returns:
         str: A message containing the token creation details.
@@ -50,7 +55,7 @@ def wow_create_token(wallet: Wallet, name: str, symbol: str) -> str:
             args={
                 "_tokenCreator": wallet.default_address.address_id,
                 "_platformReferrer": "0x0000000000000000000000000000000000000000",
-                "_tokenURI": GENERIC_TOKEN_METADATA_URI,
+                "_tokenURI": token_uri or GENERIC_TOKEN_METADATA_URI,
                 "_name": name,
                 "_symbol": symbol,
             },